+++ package.json
@@ -7,7 +7,7 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "export":"next export",
+    "export": "next export",
     "post-update": "echo \"codesandbox preview only, need an update\" && yarn upgrade --latest"
   },
   "dependencies": {
@@ -22,12 +22,15 @@
     "@mui/x-data-grid": "^5.17.24",
     "@next/font": "latest",
     "@reduxjs/toolkit": "^1.9.2",
+    "autoprefixer": "^10.4.13",
     "axios": "^1.3.3",
     "mui-image": "^1.0.7",
     "next": "latest",
+    "postcss": "^8.4.21",
     "react": "latest",
     "react-dom": "latest",
     "react-hook-form": "^7.43.1",
+    "react-image-gallery": "^1.2.11",
     "react-redux": "^8.0.5",
     "yup": "^1.0.0"
   },
@@ -37,6 +40,7 @@
     "@types/react": "latest",
     "eslint": "latest",
     "eslint-config-next": "latest",
+    "tailwindcss": "^3.2.7",
     "typescript": "latest"
   }
 }
