--- pages/reset-password.tsx
@@ -1,158 +1,0 @@
-import Box from "@mui/material/Box";
-import TextField from "@mui/material/TextField";
-import Button from "@mui/material/Button";
-import Grid from "@mui/material/Grid";
-import React, { useEffect, useState } from "react";
-import Avatar from "@mui/material/Avatar";
-import CssBaseline from "@mui/material/CssBaseline";
-import Link from "@mui/material/Link";
-import LockOutlinedIcon from "@mui/icons-material/LockOutlined";
-import Typography from "@mui/material/Typography";
-import Container from "@mui/material/Container";
-import { createTheme, ThemeProvider } from "@mui/material/styles";
-import { useForm, SubmitHandler } from "react-hook-form";
-import { useRouter } from "next/router";
-import Router from "next/router";
-import axiosRequest from "../src/utils/axiosRequest";
-
-//Copyright function at the bottom of the page
-function Copyright(props: any) {
-  return (
-    <Typography
-      variant="body2"
-      color="text.secondary"
-      align="center"
-      {...props}
-    >
-      {"Copyright © "}
-      <Link color="inherit" href="">
-        FotoPie
-      </Link>{" "}
-      {new Date().getFullYear()}
-      {"."}
-    </Typography>
-  );
-}
-
-interface FormData {
-  password: string;
-}
-
-const theme = createTheme();
-
-function ResetPassword() {
-  const [password, setPassword] = useState("");
-
-  const handleEmailChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setPassword(event.target.value);
-  };
-
-  const router = useRouter();
-  const { token } = router.query;
-
-  // const searchParams = new URLSearchParams(window.location.search);
-  // const token = searchParams.get("token");
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    try {
-      const response = await fetch(
-        `http://localhost:9090/api/reset/resetPassword?token=${token}`,
-        {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({
-            password,
-          }),
-        }
-      );
-
-      const data = await response.json();
-
-      //redirecting page based on server response code
-      switch (response.status) {
-        case 200:
-          Router.push("/login");
-          break;
-        case 401:
-          Router.push("/invalid token");
-          break;
-      }
-      console.log(data);
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  const {
-    register,
-    formState: { errors },
-  } = useForm<FormData>();
-
-  return (
-    <ThemeProvider theme={theme}>
-      <Container component="main" maxWidth="xs">
-        <CssBaseline />
-        <Box
-          sx={{
-            marginTop: 8,
-            display: "flex",
-            flexDirection: "column",
-            alignItems: "center",
-          }}
-        >
-          <Avatar sx={{ m: 1, bgcolor: "secondary.main" }}>
-            <LockOutlinedIcon />
-          </Avatar>
-          <Typography component="h1" variant="h5">
-            Create New Password
-          </Typography>
-          <Box component="form" onSubmit={handleSubmit} sx={{ mt: 3 }}>
-            <Grid container spacing={2}>
-              <Grid item xs={12}>
-                <TextField
-                  required
-                  fullWidth
-                  label="Password"
-                  type="password"
-                  id="password"
-                  autoComplete="new-password"
-                  value={password}
-                  onChange={handleEmailChange}
-                  // {...register("password", {
-                  //   required: true,
-                  //   minLength: 8,
-                  //   maxLength: 20,
-                  //   pattern: {
-                  //     value: /^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{8,}$/,
-                  //     message:
-                  //       "Password must be at least 8 characters long and contain at least one letter and one number",
-                  //   },
-                  // })}
-                  // error={!!errors.password}
-                  // helperText={errors.password?.message}
-                />
-              </Grid>
-            </Grid>
-            <Button
-              type="submit"
-              fullWidth
-              variant="contained"
-              sx={{ mt: 3, mb: 2 }}
-            >
-              Submit
-            </Button>
-            <Grid container justifyContent="flex-end">
-              <Grid item>
-                <p>Please create a new password and keep it safely</p>
-              </Grid>
-            </Grid>
-          </Box>
-        </Box>
-        <Copyright sx={{ mt: 5 }} />
-      </Container>
-    </ThemeProvider>
-  );
-}
-
-export default ResetPassword;
