+++ pages/reset-password/reset-request.tsx
@@ -1,0 +1,146 @@
+import Box from "@mui/material/Box";
+import TextField from "@mui/material/TextField";
+import Button from "@mui/material/Button";
+import Grid from "@mui/material/Grid";
+import React, { useEffect, useState } from "react";
+import Avatar from "@mui/material/Avatar";
+import CssBaseline from "@mui/material/CssBaseline";
+import Link from "@mui/material/Link";
+import LockOutlinedIcon from "@mui/icons-material/LockOutlined";
+import Typography from "@mui/material/Typography";
+import Container from "@mui/material/Container";
+import { createTheme, ThemeProvider } from "@mui/material/styles";
+import { useForm, SubmitHandler } from "react-hook-form";
+import Router from "next/router";
+
+//Copyright function at the bottom of the page
+function Copyright(props: any) {
+  return (
+    <Typography
+      variant="body2"
+      color="text.secondary"
+      align="center"
+      {...props}
+    >
+      {"Copyright © "}
+      <Link color="inherit" href="">
+        FotoPie
+      </Link>{" "}
+      {new Date().getFullYear()}
+      {"."}
+    </Typography>
+  );
+}
+
+interface FormData {
+  email: string;
+}
+
+const theme = createTheme();
+
+//main function
+function ResetRequest() {
+  const [email, setEmail] = useState("");
+
+  const handleEmailChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setEmail(event.target.value);
+  };
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    try {
+      const response = await fetch(
+        "http://localhost:9090/api/reset/resetRequest",
+        {
+          method: "POST",
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify({
+            email,
+          }),
+        }
+      );
+      const data = await response.json();
+
+      //redirecting page based on server response code
+      switch (response.status) {
+        case 200:
+          Router.push("/email-sent");
+          break;
+        case 404:
+          Router.push("/user-not-exist");
+          break;
+      }
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
+  const {
+    register,
+    formState: { errors },
+  } = useForm<FormData>();
+
+  //Front end page
+  return (
+    <ThemeProvider theme={theme}>
+      <Container component="main" maxWidth="xs">
+        <CssBaseline />
+        <Box
+          sx={{
+            marginTop: 8,
+            display: "flex",
+            flexDirection: "column",
+            alignItems: "center",
+          }}
+        >
+          <Avatar sx={{ m: 1, bgcolor: "secondary.main" }}>
+            <LockOutlinedIcon />
+          </Avatar>
+          <Typography component="h1" variant="h5">
+            Reset Password
+          </Typography>
+          <Box component="form" onSubmit={handleSubmit} sx={{ mt: 3 }}>
+            <Grid container spacing={2}>
+              <Grid item xs={12}>
+                <TextField
+                  required
+                  fullWidth
+                  id="email"
+                  label="Email Address"
+                  autoComplete="email"
+                  value={email}
+                  onChange={handleEmailChange}
+                  // {...register("email", {
+                  //   required: true,
+                  //   pattern: {
+                  //     value: /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}$/i,
+                  //     message: "Invalid email address",
+                  //   },
+                  // })}
+                  // error={!!errors.email}
+                  // helperText={errors.email?.message}
+                />
+              </Grid>
+            </Grid>
+            <Button
+              type="submit"
+              fullWidth
+              variant="contained"
+              sx={{ mt: 3, mb: 2 }}
+            >
+              Send Verification
+            </Button>
+            <Grid container justifyContent="flex-end">
+              <Grid item>
+                <p>Please click on the link in the verification email</p>
+              </Grid>
+            </Grid>
+          </Box>
+        </Box>
+        <Copyright sx={{ mt: 5 }} />
+      </Container>
+    </ThemeProvider>
+  );
+}
+
+export default ResetRequest;
